requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-heroku==0.3.1
djangorestframework==3.10.3
geographiclib==1.50
geopy==1.20.0
gunicorn==19.9.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
